Split generic and cloudflare specific code in config generator
